# aspNetCore-SDK/src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Accounts API",
    "version" : "1.0.4",
    "title" : "Accounts API",
    "contact" : {
      "email" : "you@your-company.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/MConneely/AccountsAPI/1.0.4",
  "tags" : [ {
    "name" : "admins",
    "description" : "Secured Admin-only calls"
  }, {
    "name" : "developers",
    "description" : "Operations available to regular developers"
  } ],
  "schemes" : [ "https" ],
  "paths" : {
    "/accounts" : {
      "get" : {
        "tags" : [ "developers" ],
        "summary" : "searches accounts",
        "description" : "By passing in the appropriate options, you can search for\navailable accounts in the system\n",
        "operationId" : "searchaccounts",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "number_id",
          "in" : "query",
          "description" : "pass an optional search string for looking up accounts",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "NumberId",
          "in" : "query",
          "description" : "pass an optional search string for looking up accounts",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "skip",
          "in" : "query",
          "description" : "number of records to skip for pagination",
          "required" : false,
          "type" : "integer",
          "minimum" : 0,
          "format" : "int32"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "maximum number of records to return",
          "required" : false,
          "type" : "integer",
          "maximum" : 50,
          "minimum" : 0,
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/AccountsItem"
              }
            }
          },
          "201" : {
            "description" : "searcsdfsdfsdfh results matfsddsfching criteria",
            "schema" : {
              "type" : "string",
              "format" : "Regex"
            }
          },
          "400" : {
            "description" : "bad input parameter"
          }
        }
      },
      "post" : {
        "tags" : [ "admins" ],
        "summary" : "adds an accounts item",
        "description" : "Adds an item to the system",
        "operationId" : "addaccounts",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "accountsItem",
          "description" : "accounts item to add",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/AccountsItem"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "item created",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/AccountsItem"
              }
            }
          },
          "400" : {
            "description" : "invalid input, object invalid"
          },
          "409" : {
            "description" : "an existing item already exists"
          }
        }
      }
    },
    "/accounts/{id}" : {
      "get" : {
        "summary" : "Find account by ID",
        "description" : "Returns a single account",
        "operationId" : "getAccountByID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of account to get",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/AccountsItem"
            }
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Product not found"
          }
        }
      }
    }
  },
  "definitions" : {
    "AccountsItem" : {
      "type" : "object",
      "required" : [ "accountType", "createdDate", "id", "name" ],
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid",
          "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "name" : {
          "type" : "string",
          "example" : "Joe Bloggs"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "2016-08-29T09:12:33.001Z"
        },
        "accountType" : {
          "$ref" : "#/definitions/AccountTypeModel"
        }
      },
      "example" : {
        "createdDate" : "2016-08-29T09:12:33.001Z",
        "accountType" : {
          "phone" : "408-867-5309",
          "name" : "Personal",
          "homePage" : "https://www.personal-accounts.com"
        },
        "name" : "Joe Bloggs",
        "id" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
      }
    },
    "AccountsItemModel" : {
      "type" : "object",
      "required" : [ "accountType", "createdDate", "id", "name" ],
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid",
          "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "name" : {
          "type" : "string",
          "example" : "Joe Bloggs"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "2016-08-29T09:12:33.001Z"
        },
        "accountType" : {
          "$ref" : "#/definitions/AccountTypeModel"
        }
      },
      "example" : {
        "createdDate" : "2016-08-29T09:12:33.001Z",
        "accountType" : {
          "phone" : "408-867-5309",
          "name" : "Personal",
          "homePage" : "https://www.personal-accounts.com"
        },
        "name" : "Joe Bloggs",
        "id" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
      }
    },
    "AccountTypeModel" : {
      "required" : [ "name" ],
      "properties" : {
        "name" : {
          "type" : "string",
          "example" : "Personal"
        },
        "homePage" : {
          "type" : "string",
          "format" : "url",
          "example" : "https://www.personal-accounts.com"
        },
        "phone" : {
          "type" : "string",
          "example" : "408-867-5309"
        }
      },
      "example" : {
        "phone" : "408-867-5309",
        "name" : "Personal",
        "homePage" : "https://www.personal-accounts.com"
      }
    }
  },
  "x-filter" : {
    "review-status" : "Pending"
  },
  "x-environment" : "STAGING"
}